backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-docx==0.8.11
pdfplumber==0.10.3
Pillow==10.0.0
pytesseract==0.3.10
nltk==3.8.1
requests==2.31.0
